M4 day 10 submissions
M4/runner.sh does not yet understand the grid notation accepted by
Python's runner, but even if it did, today's M4 golf doesn't quite
produce that format (@ instead of #, wrong whitespace).  Thus this is
not counted as a valid score.  Still, it may provide ideas, so it is
worth documenting.