M4 day 21 submission
M4 lacks 64-bit numbers, so I have to use the shell. I think limiting
things to just $(()) integer math in the shell is fair, which scores
348 (or 347 if there is a trailing newline in the input).  But once
the shell is involved, I was able to further squeeze things to 226
bytes by depending on bash and GNU bc for an analytical approach with
fractional math, rather than an undo stack in m4 integral math - but
that is far enough outside the realm of pure GNU m4 to not count it as
the official scoreboard size.